Issue 88 - Give multiple entry points a shot.
