+++ lib/mentionsPlugin.js
@@ -2,6 +2,8 @@
  * Plugin to parse `@user` mentions in habitica and turn them into renderable tokens.
  */
 
+const isLinkOrEmail = require('./util/isLinkOrEmail');
+
 const mentionRegex = /@[\w-]+/;
 const whiteSpaceRegex = /\s+/;
 
@@ -15,7 +17,7 @@ function partOfLinkOrEmail (pretext, text, md) {
   const prefix = preTextWords[preTextWords.length - 1];
   const possibleLink = prefix + atWord;
 
-  return md.isLinkOrEmail(possibleLink);
+  return isLinkOrEmail(possibleLink, md.linkify);
 }
 
 /**
