+++ lib/unsafe.js
@@ -1,0 +1,5 @@
+const createMdInstance = require('./util/createMdInstance');
+
+module.exports = createMdInstance({
+  html: true,
+});
