+++ lib/withMentions.js
@@ -1,0 +1,4 @@
+const createMdInstance = require('./util/createMdInstance');
+const mentionsPlugin = require('./mentionsPlugin');
+
+module.exports = createMdInstance().use(mentionsPlugin);
